requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
gunicorn==21.2.0 